+++ shm/ship.cpp
@@ -4,16 +4,20 @@
 
 Ship::Ship() : id_(-1){};
 
-Ship::Ship(uint8_t capacity, uint8_t maxCrew, uint8_t speed, const std::string& name, uint8_t id)
+Ship::Ship(uint32_t capacity,
+           uint32_t maxCrew,
+           uint32_t speed,
+           const std::string& name,
+           uint32_t id)
     : capacity_(capacity), maxCrew_(maxCrew), crew_(0), speed_(speed), name_(name), id_(id){};
 
-Ship::Ship(uint8_t maxCrew, uint8_t speed, uint8_t id) : Ship(0, maxCrew, speed, "", id){};
+Ship::Ship(uint32_t maxCrew, uint32_t speed, uint32_t id) : Ship(0, maxCrew, speed, "", id){};
 
 void Ship::setName(const std::string& name) {
     name_ = name;
 }
 
-Ship& Ship::operator-=(uint8_t crew) {
+Ship& Ship::operator-=(uint32_t crew) {
     if (crew_ < crew) {
         crew_ -= crew;
     }
@@ -21,7 +25,7 @@ Ship& Ship::operator-=(uint8_t crew) {
     return *this;
 }
 
-Ship& Ship::operator+=(uint8_t crew) {
+Ship& Ship::operator+=(uint32_t crew) {
     if (crew_ + crew <= maxCrew_) {
         crew_ += crew;
     }
@@ -29,7 +33,7 @@ Ship& Ship::operator+=(uint8_t crew) {
     return *this;
 }
 
-uint8_t Ship::getId() const {
+uint32_t Ship::getId() const {
     return id_;
 }
 
@@ -37,14 +41,14 @@ std::string Ship::getName() const {
     return name_;
 }
 
-uint8_t Ship::getSpeed() const {
+uint32_t Ship::getSpeed() const {
     return speed_;
 }
 
-uint8_t Ship::getMaxCrew() const {
+uint32_t Ship::getMaxCrew() const {
     return maxCrew_;
 }
 
-uint8_t Ship::getCapacity() const {
+uint32_t Ship::getCapacity() const {
     return capacity_;
 }
